requirement.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==2.2.4
numpy==1.20.2
Pillow==8.2.0
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
